streamlit_app/requirements.txt
streamlit
paho-mqtt
plotly
pandas
streamlit-autorefresh
streamlit-mermaid
